+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG Testimonials Widget
 
 ## master
+* [Auto-suggest category and tag options](http://wordpress.org/support/topic/feature-request-menus-in-the-widget)
 
 ## 2.15.0-beta
 * Add 2.15.0 upgrade notice